Revert "Contagem de performance nos minigames"
This reverts commit c46c9d94670b4d4f1c68b8d471cd941801e7e47a.

+++ Time-1/Assets/Game/Scripts/Stats/playerStats.cs
@@ -17,21 +17,21 @@ public class playerStats : MonoBehaviour
         maxHealth = playerBase.maxHealth;
     }
 
-    public void raiseStats(int type, float value)
+    public void raiseStats(int type, int value)
     {
         if (type == 1)
         {
-            velocity += (int)value;
-            maxEnergy += (int)value;
+            velocity += value;
+            maxEnergy += value;
         }
         else if (type == 2)
         {
-            defense += (int)value;
-            maxHealth += (int)value *3;
+            defense += value;
+            maxHealth += value*3;
         }
         else if (type == 3)
         {
-            attack += (int)value;
+            attack += value;
         }
     }
